src/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
PyYAML>=5.4.1
scikit-learn>=1.0.0
joblib>=1.1.0